+++ examples/data-sources/fmc_system_information/data-source.tf
@@ -1,0 +1,2 @@
+data "fmc_system_information" "example" {
+}