+++ internal/provider/data_source_fmc_system_information.go
@@ -1,0 +1,145 @@
+// Copyright © 2023 Cisco Systems, Inc. and its affiliates.
+// All rights reserved.
+//
+// Licensed under the Mozilla Public License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://mozilla.org/MPL/2.0/
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+// Section below is generated&owned by "gen/generator.go". //template:begin imports
+import (
+	"context"
+	"fmt"
+
+	"github.com/CiscoDevNet/terraform-provider-fmc/internal/provider/helpers"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
+	"github.com/netascode/go-fmc"
+)
+
+// End of section. //template:end imports
+
+// Section below is generated&owned by "gen/generator.go". //template:begin model
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ datasource.DataSource              = &SystemInformationDataSource{}
+	_ datasource.DataSourceWithConfigure = &SystemInformationDataSource{}
+)
+
+func NewSystemInformationDataSource() datasource.DataSource {
+	return &SystemInformationDataSource{}
+}
+
+type SystemInformationDataSource struct {
+	client *fmc.Client
+}
+
+func (d *SystemInformationDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_system_information"
+}
+
+func (d *SystemInformationDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		// This description is used by the documentation generator and the language server.
+		MarkdownDescription: helpers.NewAttributeDescription("Get FMC system information.").String,
+
+		Attributes: map[string]schema.Attribute{
+			"hostname": schema.StringAttribute{
+				MarkdownDescription: "Hostname.",
+				Computed:            true,
+			},
+			"lsp_version": schema.StringAttribute{
+				MarkdownDescription: "LSP version.",
+				Computed:            true,
+			},
+			"model": schema.StringAttribute{
+				MarkdownDescription: "Model.",
+				Computed:            true,
+			},
+			"platform": schema.StringAttribute{
+				MarkdownDescription: "Platform.",
+				Computed:            true,
+			},
+			"serial_number": schema.StringAttribute{
+				MarkdownDescription: "Serial number.",
+				Computed:            true,
+			},
+			"server_version": schema.StringAttribute{
+				MarkdownDescription: "Server version.",
+				Computed:            true,
+			},
+			"sru_version": schema.StringAttribute{
+				MarkdownDescription: "SRU version.",
+				Computed:            true,
+			},
+			"type": schema.StringAttribute{
+				MarkdownDescription: "Type of the object.",
+				Computed:            true,
+			},
+			"uptime": schema.StringAttribute{
+				MarkdownDescription: "Uptime.",
+				Computed:            true,
+			},
+			"vdb_version": schema.StringAttribute{
+				MarkdownDescription: "VDB version.",
+				Computed:            true,
+			},
+		},
+	}
+}
+
+func (d *SystemInformationDataSource) Configure(_ context.Context, req datasource.ConfigureRequest, _ *datasource.ConfigureResponse) {
+	if req.ProviderData == nil {
+		return
+	}
+
+	d.client = req.ProviderData.(*FmcProviderData).Client
+}
+
+// End of section. //template:end model
+
+// Section below is generated&owned by "gen/generator.go". //template:begin read
+
+func (d *SystemInformationDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	var config SystemInformation
+
+	// Read config
+	diags := req.Config.Get(ctx, &config)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	// Set request domain if provided
+	reqMods := [](func(*fmc.Req)){}
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", "System Information"))
+	urlPath := config.getPath()
+	res, err := d.client.Get(urlPath, reqMods...)
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object, got error: %s", err))
+		return
+	}
+
+	config.fromBody(ctx, res)
+
+	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", "System Information"))
+
+	diags = resp.State.Set(ctx, &config)
+	resp.Diagnostics.Append(diags...)
+}
+
+// End of section. //template:end read