+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - (Enh) Access Rule and Category: Add posibility to create at specific location
 - (Enh) Device interfaces data sources now accept `logical_name` as search parameter
 - (Enh) If `items` in bulk data source is empty, all objects from FMC will be loaded
+- (Enh) Add `fmc_system_information` data source
 
 ## 2.1.0